more updates on labor routes

+++ src/controllers/labor/DeleteLaborController.ts
@@ -4,9 +4,12 @@ import { Request, Response } from "express";
 
 class DeleteLaborController {
   async handle(req: Request, res: Response) {
-    const { laborId } = req.body;
+    const userId = req.params.userId;
+    if (!userId) {
+      return res.status(400).json({ error: "User ID is required" });
+    }
     const deleteLaborService = new DeleteLaborService();
-    const deletedLabor = await deleteLaborService.execute(laborId);
+    const deletedLabor = await deleteLaborService.execute(userId);
     return res.json(deletedLabor);
   }
 }
